US-31-Task-97-Created repository and fixed service issues for view my team API for coach

+++ SoccerTournamentBackend/src/main/java/com/asu/soccer/tournament/SoccerTournament/player/service/PlayerServiceImpl.java
@@ -58,12 +58,12 @@ public class PlayerServiceImpl implements PlayerService {
 		//userEntity.setReg_date(formattedDate);
 		
 		
-		Integer id = (int)(Math.random() * 100000);
+		long id = (long)(Math.random() * 100000);
 		while(userRepository.findById(id) != null)
 		{
-			id = (int)(Math.random() * 100000);
+			id = (long)(Math.random() * 100000);
 		}
-		userEntity.setId(id);
+		userEntity.setId((int)id);
 		
 		UserEntity createdUser = userRepository.save(userEntity);
 		
@@ -74,12 +74,12 @@ public class PlayerServiceImpl implements PlayerService {
 		}
 		
 		PlayerEntity playerEntity = new PlayerEntity();
-		playerEntity.setId(id); // same as the ID for userEntity
+		playerEntity.setId((int)id); // same as the ID for userEntity
 		playerEntity.setGames_played(0);
 		playerEntity.setRed_cards(0);
 		playerEntity.setSubs(0);
 		playerEntity.setTeam_details_id(null);
-		playerEntity.setUser_details_id(id);
+		playerEntity.setUser_details_id((int)id);
 	
 		PlayerEntity createdPlayer = playerRepository.save(playerEntity);
